requirements.txt
pandas>=2.2.0
openpyxl>=3.1.0
selenium>=4.10.0
requests>=2.32.0
webdriver-manager>=4.0.0
PyPDF2>=3.0.0
pillow>=10.0.0
pyautogui>=0.9.54
python-dotenv>=1.0.0

